+++ src/main/java/session/Session.java
@@ -1,4 +1,4 @@
-package model;
+package session;
 
 import java.util.UUID;
